requirements.txt
# pip install -r requirements.txt
cryptography
pyOpenSSL
certifi
msedge-selenium-tools
pyperclip
selenium
numpy